django-ninja-no-docker/requirements.txt
django 
django-ninja-extra